+++ examples/src/main/java/io/github/com/ensure4j/examples/MultipleBoolCheck.java
@@ -7,6 +7,7 @@ public class MultipleBoolCheck {
     public static void main(String[] args) {
         // Good practice is to keep one pre-condition per line.
         // It's easier to know which pre-condition failed when it fails
+        // with a well-defined error message.
         Ensure.isTrue(isFullMoon(), "no full moon");
         Ensure.isTrue(hasGarlicBread(), "no garlic bread");
     }