+++ lib/src/main/java/io/github/mangila/ensure4j/package-info.java
@@ -1,5 +1,12 @@
 /**
- * Ensure your Java code
+ * <p>
+ * Annotated with {@link org.jspecify.annotations.NullUnmarked} since this package handles Unspecified nullness
+ * </p>
+ * Package contains:
+ * <ul>
+ *     <li>{@link io.github.mangila.ensure4j.Ensure} - The utility implementation</li>
+ *     <li>{@link io.github.mangila.ensure4j.EnsureException} - The exception that will be thrown when a precondition fails</li>
+ * </ul>
  */
 @NullUnmarked
 package io.github.mangila.ensure4j;