+++ lib/src/test/java/io/github/mangila/ensure4j/EnsureTest.java
@@ -40,7 +40,7 @@ public class EnsureTest {
     }
 
     @Test
-    void testMethodSize() {
+    void archTest() {
         String javaPackage = "io.github.mangila.ensure4j";
         ArchRuleDefinition.classes()
                 .that()
@@ -58,8 +58,8 @@ public class EnsureTest {
                                 .map(JavaMember::getName)
                                 .toList();
                         assertPublicMethods(publicMethods);
-                        int size = item.getMethods().size();
-                        assertThat(size).isEqualTo(36);
+                        int totalMethods = item.getMethods().size();
+                        assertThat(totalMethods).isEqualTo(39);
                     }
 
                     /**
@@ -68,7 +68,7 @@ public class EnsureTest {
                     private void assertPublicMethods(List<String> publicMethodNames) {
                         int distinctPublicMethodCount = publicMethodNames.size();
                         assertThat(distinctPublicMethodCount)
-                                .isEqualTo(34);
+                                .isEqualTo(37);
                         Map<String, Long> counts = publicMethodNames.stream()
                                 .collect(Collectors.groupingBy(
                                         methodName -> methodName,
@@ -87,6 +87,7 @@ public class EnsureTest {
                                 case "notBlank" -> assertThat(counts.get(methodName)).isEqualTo(3L);
                                 case "notBlankOrElse" -> assertThat(counts.get(methodName)).isEqualTo(1L);
                                 case "notEmpty" -> assertThat(counts.get(methodName)).isEqualTo(9L);
+                                case "notContainsNull" -> assertThat(counts.get(methodName)).isEqualTo(3L);
                                 case "equals" -> assertThat(counts.get(methodName)).isEqualTo(3L);
                                 default -> throw new IllegalStateException("Unexpected value: " + methodName);
                             }