+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 0.0.6
+
+Prettier Javadocs
+
+Added methods:
+
+- notContainsNull(Collection)
+
 ## 0.0.5
 
 More methods.